chore(docs): Update Boost Installation
We recommended that people use `brew` to install boost. Let's expand
on that, and note that people should use the package manager of their
choice, assuming it has a recent enough version of boost.

Also add a link to the Boost SourceForge documentation, and the name
of the bundled library for apt users.