feat(analyzer): Check for duplicate impls by ingot
Previously, when looking for duplicate implementations of a trait, we
checked at the module level. However this meant that nothing stopped
you from adding duplicate impls in two different modules in an
ingot. Let's move this check a layer up to ensure that we can see
avoid adding two impls for a specific trait.